+++ packages/observable-client/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./getObservableClient"
+export { isBestOrFinalizedBlock } from "./chainHead/streams/block-operations"
 export {
   BlockNotPinnedError,
   BlockPrunedError,