+++ packages/observable-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+# Added
+
+- expose `isBestOrFinalizedBlock` helper
+
 ## 0.6.5 - 2024-12-17
 
 ### Fixed